remove survey
took survey down from all pages except the "characters" page. felt the survey  should be unique to characters page as it will have more information on the characters the survey will be about. althought most characters are locked right now, I felt this was a good way to give readers incentive to read ahead and find out who the other characters are as the story will progress a little bit ahead of the updates on the website.